refactor: index.d.ts와 index.ts가 상충되는 오류를 수정
index.ts -> general.ts로 변경한다
