+++ src/pages/AddCardAlias.tsx
@@ -5,7 +5,7 @@ import CardAliasInput from '../components/@common/CardAliasInput';
 import React from 'react';
 import SubmitButton from '../components/@common/SubmitButton';
 import { useNavigate } from 'react-router-dom';
-import { CardList } from '../types/';
+import { CardList } from '../types/general';
 import AddCardLoadingSpinner from '../components/AddCardLoadingSpinner/AddCardLoadingSpinner';
 import { Link } from 'react-router-dom';
 import useAddCardAlias from '../hooks/useAddCardAlias';
